requirements.txt
seldom==3.7.0
#xlrd==1.2.0
loguru~=0.6.0
requests~=2.32.3
urllib3~=2.2.2
openpyxl~=3.1.5
ddt~=1.7.2
aiomysql~=0.2.0
BeautifulReport~=0.1.3
HTMLReport~=2.4.0
XTestRunner~=1.7.4
python-dotenv~=1.0.1
openai~=1.57.0
bs4~=0.0.2
beautifulsoup4~=4.12.3
aiofiles~=24.1.0
pyppeteer~=2.0.0
locust~=2.31.4
opencv-python~=4.10.0.84
numpy~=2.1.3
playwright~=1.48.0
pycryptodome~=3.20.0
fake-useragent~=1.5.1
sympy~=1.13.1
psutil~=6.0.0
DrissionPage~=4.1.0.11
lxml~=5.3.0
allure-python-commons~=2.13.5
allure-pytest~=2.13.5
jsonpath-ng~=1.6.1
pytest~=8.3.4
#baidu-aip~=4.16.13
#chardet~=5.2.0
pandas~=2.2.3
python-dateutil~=2.9.0.post0